feat: Add Manchester Derby jersey to homepage
This commit adds the Manchester Derby jersey to the homepage of the website. The jersey is fetched from the `manchester_jercey` data file and displayed in the `HomeProductSection` component. This addition enhances the variety of products available on the homepage and provides more options for users to explore and purchase.